Provision from image name as well as image id
