+++ src/test/java/jenkins/plugins/openstack/compute/JCloudsCloudTest.java
@@ -193,138 +193,6 @@ public class JCloudsCloudTest {
         assertEquals(original.getRawSlaveOptions(), JCloudsCloud.getByName("openstack").getRawSlaveOptions());
     }
 
-    @Test @SuppressWarnings("deprecation")
-    public void manullyProvisionAndKill() throws Exception {
-        Computer computer = j.provisionDummySlave("label").toComputer();
-        assertTrue("Slave should be connected", computer.isOnline());
-
-        computer.doDoDelete();
-        assertTrue("Slave is temporarily offline", computer.isTemporarilyOffline());
-
-        j.triggerOpenstackSlaveCleanup();
-        assertEquals("Slave is discarded", null, j.jenkins.getComputer("provisioned"));
-    }
-
-    @Test
-    public void provisionSlaveOnDemand() throws Exception {
-        JCloudsCloud cloud = j.createCloudProvisioningDummySlaves("label");
-
-        Computer[] originalComputers = j.jenkins.getComputers();
-
-        j.jenkins.setNumExecutors(0);
-        FreeStyleProject p = j.createFreeStyleProject();
-        // Provision with label
-        p.setAssignedLabel(Label.get("label"));
-        Node node = j.buildAndAssertSuccess(p).getBuiltOn();
-        assertThat(node, Matchers.instanceOf(JCloudsSlave.class));
-        node.toComputer().doDoDelete();
-        j.triggerOpenstackSlaveCleanup();
-        assertThat(originalComputers, arrayContainingInAnyOrder(j.jenkins.getComputers()));
-
-        // Provision without label
-        p.setAssignedLabel(null);
-        assertThat(j.buildAndAssertSuccess(p).getBuiltOn(), Matchers.instanceOf(JCloudsSlave.class));
-
-        Openstack os = cloud.getOpenstack();
-        verify(os, atLeastOnce()).getRunningNodes();
-        verify(os, times(2)).bootAndWaitActive(any(ServerCreateBuilder.class), any(Integer.class));
-        verify(os, times(2)).assignFloatingIp(any(Server.class), eq("custom"));
-        verify(os, times(2)).updateInfo(any(Server.class));
-        verify(os, atLeastOnce()).destroyServer(any(Server.class));
-
-        verifyNoMoreInteractions(os);
-    }
-
-    @Test
-    public void doNotProvisionOnceInstanceCapReached() {
-        JCloudsSlaveTemplate restricted = j.dummySlaveTemplate(SlaveOptions.builder().instanceCap(1).build(), "restricted");
-        JCloudsSlaveTemplate open = j.dummySlaveTemplate(SlaveOptions.builder().instanceCap(null).build(), "open");
-        JCloudsCloud cloud = j.dummyCloud(SlaveOptions.builder().instanceCap(4).build(), restricted, open);
-
-        Server restrictedMachine = j.mockServer().metadataItem(JCloudsSlaveTemplate.OPENSTACK_TEMPLATE_NAME_KEY, restricted.name).get();
-        Server openMachine = j.mockServer().metadataItem(JCloudsSlaveTemplate.OPENSTACK_TEMPLATE_NAME_KEY, open.name).get();
-
-        List<Server> running = new ArrayList<>();
-
-        when(cloud.getOpenstack().getRunningNodes()).thenReturn(running);
-
-        // Template quota exceeded
-        assertEquals(1, cloud.provision(Label.get("restricted"), 2).size());
-        running.add(restrictedMachine);
-        assertEquals(0, cloud.provision(Label.get("restricted"), 1).size());
-        assertEquals(1, restricted.getRunningNodes().size());
-
-        // Cloud quota exceeded
-        assertEquals(2, cloud.provision(Label.get("open"), 2).size());
-        running.add(openMachine);
-        running.add(openMachine);
-        assertEquals(2, open.getRunningNodes().size());
-        assertEquals(1, cloud.provision(Label.get("open"), 2).size());
-        running.add(openMachine);
-        assertEquals(3, open.getRunningNodes().size());
-
-        assertEquals(0, cloud.provision(Label.get("restricted"), 1).size());
-        assertEquals(0, cloud.provision(Label.get("open"), 1).size());
-    }
-
-    @Test @Issue("https://github.com/jenkinsci/openstack-cloud-plugin/issues/31")
-    public void abortProvisioningWhenOpenstackFails() throws Exception {
-        JCloudsSlaveTemplate template = j.dummySlaveTemplate("label");
-        JCloudsCloud cloud = j.dummyCloud(template);
-        Openstack os = cloud.getOpenstack();
-        when(os.bootAndWaitActive(any(ServerCreateBuilder.class), any(Integer.class))).thenThrow(new Openstack.ActionFailed("It is broken, alright!"));
-
-        FreeStyleProject p = j.createFreeStyleProject();
-        p.setAssignedLabel(Label.get("label"));
-        Future<FreeStyleBuild> started = p.scheduleBuild2(0).getStartCondition();
-
-        Thread.sleep(1000);
-        assertFalse(started.isDone());
-
-        verify(os, atLeastOnce()).bootAndWaitActive(any(ServerCreateBuilder.class), any(Integer.class));
-    }
-
-    @Test @Issue("https://github.com/jenkinsci/openstack-cloud-plugin/issues/31")
-    public void failToProvisionManuallyWhenOpenstackFails() throws Exception {
-        JCloudsSlaveTemplate template = j.dummySlaveTemplate("label");
-        final JCloudsCloud cloud = j.dummyCloud(template);
-        Openstack os = cloud.getOpenstack();
-        when(os.bootAndWaitActive(any(ServerCreateBuilder.class), any(Integer.class))).thenThrow(new Openstack.ActionFailed("It is broken, alright!"));
-
-        JenkinsRule.WebClient wc = j.createWebClient();
-        wc.setThrowExceptionOnFailingStatusCode(false);
-        wc.setPrintContentOnFailingStatusCode(false);
-        Page page = wc.getPage(wc.addCrumb(new WebRequestSettings(
-                new URL(wc.getContextPath() + "cloud/openstack/provision?name=" + template.name),
-                HttpMethod.POST
-        )));
-
-        assertThat(page.getWebResponse().getContentAsString(), containsString("It is broken, alright!"));
-
-        verify(os, times(1)).bootAndWaitActive(any(ServerCreateBuilder.class), any(Integer.class));
-    }
-
-    @Test @Issue("https://github.com/jenkinsci/openstack-cloud-plugin/issues/37")
-    public void detectBootTimingOut() {
-        JCloudsSlaveTemplate template = j.dummySlaveTemplate("label");
-        final JCloudsCloud cloud = j.dummyCloud(template);
-        Openstack os = cloud.getOpenstack();
-        Server server = j.mockServer().name("provisioned").status(Server.Status.BUILD).get();
-        when(os.bootAndWaitActive(any(ServerCreateBuilder.class), any(Integer.class))).thenCallRealMethod();
-        when(os._bootAndWaitActive(any(ServerCreateBuilder.class), any(Integer.class))).thenReturn(server);
-        when(os.updateInfo(eq(server))).thenReturn(server);
-
-        try {
-            template.provision(cloud);
-            fail();
-        } catch (Openstack.ActionFailed ex) {
-            assertThat(ex.getMessage(), containsString("Failed to boot server in time"));
-            assertThat(ex.getMessage(), containsString("status=BUILD"));
-        }
-
-        verify(os).destroyServer(eq(server));
-    }
-
     @Test @LocalData
     public void globalConfigMigrationFromV1() throws Exception {
         JCloudsCloud cloud = (JCloudsCloud) j.jenkins.getCloud("OSCloud");
@@ -361,27 +229,5 @@ public class JCloudsCloudTest {
 
     private String fileAsString(String filename) throws IOException {
         return IOUtils.toString(getClass().getResourceAsStream(getClass().getSimpleName() + "/" + filename));
-    }
-
-    @Test
-    public void verifyOptionsPropagatedToLauncher() throws Exception {
-        SlaveOptions expected = j.dummySlaveOptions().getBuilder().slaveType(JCloudsCloud.SlaveType.SSH).retentionTime(10).build();
-        JCloudsCloud cloud = j.configureSlaveProvisioning(j.dummyCloud(
-                expected,
-                j.dummySlaveTemplate("label"),
-                j.dummySlaveTemplate(expected.getBuilder().retentionTime(42).build(), "retention")
-        ));
-
-        JCloudsSlave slave = j.provision(cloud, "label");
-
-        SSHLauncher launcher = (SSHLauncher) ((JCloudsLauncher) slave.getLauncher()).lastLauncher;
-        assertEquals(slave.getPublicAddress(), launcher.getHost());
-        assertEquals(expected.getCredentialsId(), launcher.getCredentialsId());
-        assertEquals(expected.getJvmOptions(), launcher.getJvmOptions());
-        assertEquals(10, ((JCloudsComputer) slave.toComputer()).getRetentionTime());
-
-        slave = j.provision(cloud, "retention");
-
-        assertEquals(42, ((JCloudsComputer) slave.toComputer()).getRetentionTime());
     }
 }
